respeaker_ros: add setup.py, enable 'from respaker_ros import PARAMETERS'
